Add descriptive title to docs and annotation panel
